+++ qbraid/programs/loader.py
@@ -10,8 +10,7 @@
 
 """
 Module containing top-level qbraid program loader functionality
-utilizing entrypoints via ``importlib`` for Python >= 3.10 or
-``pkg_resources`` for Python 3.9.
+utilizing entrypoints via ``importlib``.
 
 """
 from __future__ import annotations